+++ tests/test_content_type.py
@@ -21,59 +21,16 @@ def content_type_service(responses: RequestsMock):
         json={
             "swagger": "2.0",
             "paths": {
-                "/msgpackpandas": {
-                    "get": {
-                        "operationId": "get_msgpackpandas",
-                        "responses": {200: {"description": "successful operation"}},
-                        "produces": ["application/msgpackpandas"],
-                    }
-                },
                 "/json": {
                     "get": {
                         "operationId": "get_json",
                         "responses": {200: {"description": "successful operation"}},
                         "produces": ["application/json"],
                     }
-                },
+                }
             },
         },
         match_querystring=True,
-    )
-
-
-def test_msgpackpandas_content_type_with_pandas(
-    responses: RequestsMock, content_type_service
-):
-    pyxelrestgenerator = loader.load("content_type_service.yml")
-    responses.add(
-        responses.GET,
-        url="http://localhost:8956/msgpackpandas",
-        json={},
-        match_querystring=True,
-    )
-
-    assert pyxelrestgenerator.content_type_get_msgpackpandas() == [[""]]
-    assert (
-        _get_request(responses, "http://localhost:8956/msgpackpandas").headers["Accept"]
-        == "application/msgpackpandas"
-    )
-
-
-def test_msgpackpandas_content_type_without_pandas(
-    responses: RequestsMock, content_type_service
-):
-    pyxelrestgenerator = loader.load("content_type_service.yml")
-    responses.add(
-        responses.GET,
-        url="http://localhost:8956/msgpackpandas",
-        json={},
-        match_querystring=True,
-    )
-
-    assert pyxelrestgenerator.content_type_get_msgpackpandas() == [[""]]
-    assert (
-        _get_request(responses, "http://localhost:8956/msgpackpandas").headers["Accept"]
-        == "*/*"
     )